requirements.txt
pygame-ce
PyYAML
annotated-types